Remove test_api_notification.py
